# [3.x] FTI - Optimize `SceneTree` traversal