@lawnjelly 2025-05-24
@Calinou if you get a moment would you be able to look at this too?

It's a duplicate of the 4.x PR (aside from naming changes `Spatial` / `Node3D`).